chore: 연료주입 미션 패키지 통합
