+++ src/test/java/rentcar/abstractcar/AvanteTest.java
@@ -1,4 +1,4 @@
-package abstractcar;
+package rentcar.abstractcar;
 
 import static org.assertj.core.api.Assertions.assertThat;
