+++ src/test/java/rentcar/abstractcar/K5Test.java
@@ -1,4 +1,4 @@
-package abstractcar;
+package rentcar.abstractcar;
 
 import static org.assertj.core.api.Assertions.assertThat;
