+++ src/test/java/rentcar/abstractcar/SonataTest.java
@@ -1,4 +1,4 @@
-package interfacecar;
+package rentcar.abstractcar;
 
 import static org.assertj.core.api.Assertions.assertThat;
