+++ src/test/java/rentcar/interfacecar/SonataTest.java
@@ -1,4 +1,4 @@
-package abstractcar;
+package rentcar.interfacecar;
 
 import static org.assertj.core.api.Assertions.assertThat;
